split quant refs into quant_ref_utils; trim comments
